[#1] deleted double files and renamed some folders
