--- src/config/themes.ts
@@ -1,8 +1,0 @@
-﻿import type { CountdownUnit } from '../components/main/Countdown.tsx';
-
-export const defaultCountdownTheme: Record<CountdownUnit, string> = {
-    days: '#1abc9c',
-    hours: '#2980b9',
-    minutes: '#8e44ad',
-    seconds: '#f39c12',
-};
